+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/cache/NotifySendRecordCacheImpl.java
@@ -6,6 +6,7 @@ import com.dwarfeng.notify.stack.bean.key.NotifySendRecordKey;
 import com.dwarfeng.notify.stack.cache.NotifySendRecordCache;
 import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
@@ -61,35 +62,36 @@ public class NotifySendRecordCacheImpl implements NotifySendRecordCache {
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public boolean allExists(List<NotifySendRecordKey> keys) throws CacheException {
+    public boolean allExists(@SkipRecord List<NotifySendRecordKey> keys) throws CacheException {
         return delegate.allExists(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public boolean nonExists(List<NotifySendRecordKey> keys) throws CacheException {
+    public boolean nonExists(@SkipRecord List<NotifySendRecordKey> keys) throws CacheException {
         return delegate.nonExists(keys);
     }
 
     @Override
     @BehaviorAnalyse
+    @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-    public List<NotifySendRecord> batchGet(List<NotifySendRecordKey> keys) throws CacheException {
+    public List<NotifySendRecord> batchGet(@SkipRecord List<NotifySendRecordKey> keys) throws CacheException {
         return delegate.batchGet(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void batchPush(List<NotifySendRecord> entities, long timeout) throws CacheException {
+    public void batchPush(@SkipRecord List<NotifySendRecord> entities, long timeout) throws CacheException {
         delegate.batchPush(entities, timeout);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public void batchDelete(List<NotifySendRecordKey> keys) throws CacheException {
+    public void batchDelete(@SkipRecord List<NotifySendRecordKey> keys) throws CacheException {
         delegate.batchDelete(keys);
     }
 }